MF-464: Explictly remove cache keys that come from InvalidResponseCodeException
This involves some changes to cache management since it used to live on a lower
scope.
